+++ packages/contracts/src/spaces/facets/account/AppAccount.sol
@@ -68,7 +68,7 @@ contract AppAccount is IAppAccount, AppAccountBase, ReentrancyGuard, Facet {
     }
 
     /// @inheritdoc IAppAccount
-    function getAppExpiration(address app) external view returns (uint64) {
+    function getAppExpiration(address app) external view returns (uint48) {
         return _getGroupExpiration(_getInstalledAppId(app));
     }
